refactored view names
