+++ src/App.svelte
@@ -1,18 +1,21 @@
 <!-- App.svelte -->
 <script>
-  import { Router, Route, Link } from "svelte-routing";
-  import Overview from "./views/Overview.svelte";
-  import IdeaDetail from "./views/IdeaDetail.svelte";
-  import TutorialView from "./views/TutorialView.svelte";
+  import { Router, Route } from "svelte-routing";
+
+  import Home from "./views/Home.svelte";
+  import Tutorial from "./views/Tutorial.svelte";
+
+  import Profile from "./views/Profile.svelte";
+  import EditProfile from "./views/EditProfile.svelte";
+
+  import Idea from "./views/Idea.svelte";
   import PostIdea from "./views/PostIdea.svelte";
-  //import Login from "./views/Login.svelte";
-  import EditProfileView from "./views/EditProfileView.svelte";
-  import ProfileView from "./views/ProfileView.svelte";
-  import PreView from "./views/IdeaPreview.svelte";
+  import IdeaPreview from "./views/IdeaPreview.svelte";
+
+  import Job from "./views/Job.svelte";
   import PostJob from "./views/PostJob.svelte";
-  import JobView from "./views/JobView.svelte";
-  import JobManager from "./views/JobManager.svelte";
   import JobMarket from "./views/JobMarket.svelte";
+  import JobManager from "./views/JobManager.svelte";
 </script>
 
 <svelte:head>
@@ -28,19 +31,25 @@
       <!-- Navigation links here -->
     </nav>
     <main>
-      <Route path="/" component={Overview} />
-      <Route path="/overview" component={Overview} />
-      <Route path="/preview" component={PreView} />
-      <Route path="/jobmanager" component={JobManager} />
-      <Route path="/jobmarket" component={JobMarket} />
-      <Route path="/overview/:category" let:params component={Overview} />
-      <Route path="/idea/:id" component={IdeaDetail} />
-      <Route path="/job/:id" component={JobView} />
-      <Route path="/tutorial/:id" component={TutorialView} />
+      <Route path="/" component={Home} />
+      <Route path="/tutorial/:id" component={Tutorial} />
+      <Route path="/overview/:category" let:params component={Home} />
+
+      <Route path="/profile/:profile_id" let:params component={Profile} />
+      <Route
+        path="/edit_profile/:profile_id"
+        let:params
+        component={EditProfile}
+      />
+
+      <Route path="/idea/:id" component={Idea} />
       <Route path="/postidea" component={PostIdea} />
-      <Route path="/edit_profile/:profile_id" let:params component={EditProfileView} />
-      <Route path="/profile/:profile_id" let:params component={ProfileView} />
+      <Route path="/preview" component={IdeaPreview} />
+
+      <Route path="/job/:id" component={Job} />
       <Route path="/postjob/:ideaID" let:params component={PostJob} />
+      <Route path="/jobmarket" component={JobMarket} />
+      <Route path="/jobmanager" component={JobManager} />
     </main>
   </div>
 </Router>
