+++ src/views/Idea.svelte
